requirements.txt
nemo_toolkit[asr]
gradio
onnxruntime
fastapi
uvicorn
python-multipart
torchaudio
dotenv